refactor: extract BrowserBody and category utils
